Tested parser and Price. Some general bugs fixed.
Created Price and JourneyFileParser test case. Changed Price._quantity to double, implemented and tested parser save and load and solved few ArrayList attributes that didn't initialize.

+++ src/AeroGUI/src/classes/Price.java
@@ -1,26 +1,26 @@
 package classes;
 
 public class Price {
-    private Integer _quantity = 0;
+    private double _quantity = 0;
     private String _currency = "EUR";
 
     public Price() {
     }
 
-    public Price(Integer quantity) {
+    public Price(double quantity) {
         this._quantity = quantity;
     }
 
-    public Price(Integer quantity, String currency) {
+    public Price(double quantity, String currency) {
         this._quantity = quantity;
         this._currency = currency;
     }
 
-    public Integer getQuantity() {
+    public double getQuantity() {
         return _quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(double quantity) {
         this._quantity = quantity;
     }
